don't skip filling buffer if we are seeking
when we seek, the time update event is not fired because the video is
paused. because of this, if a segmentXhr is being processed, when we
seek, the player stops playing the video and does not resume. Added
logic to not skip filling the buffer when a segmentXhr is being
processed if we are seeking.