Fix EXT-X-KEY processing of relative URLs
The EXT-X-KEY URI attribute is not processed to spec.  It should be
interpreted as absolute or relative to the playlist containing the
EXT-X-KEY tag.

This is in the IETF spec
(https://tools.ietf.org/html/draft-pantos-http-live-streaming-13#section
-3.1).

The fix is simple, use the same URL resolution code the TS segments are
using.

Here is a stream that only works correctly with this fix:
  http://flashls-test.s3.amazonaws.com/7_copy.m3u8